refactor: add backup.cpp
